requirements.txt
temporalio==1.12.0
aiohttp
unstructured==0.17.2
pymilvus==2.4.8
protobuf
grpcio
python-dotenv
sentence-transformers==2.7.0
torch
